Práctica
He hecho select de la tabla entera usando GET, he insertado un usuario usando POST, he hecho la parte de actualizar usuario usando PUT y borrar un usuario usando DELETE

+++ aplicacion/src/main/java/com/eep/aplicacion/EstudianteController.java
@@ -3,6 +3,7 @@ package com.eep.aplicacion;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -10,31 +11,72 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+import java.util.Optional;
 @RestController
 @RequestMapping("/api_estudiantes")
 public class EstudianteController {
-	@Autowired
-	private EstudianteService estudianteService;
+    @Autowired
+    private EstudianteService estudianteService;
 
-	@GetMapping
-	public List<Estudiante> getAllEstudiantes() {
-		return estudianteService.getAllEstudiantes();
-	}
+    // Obtener todos los estudiantes
+    @GetMapping
+    public ResponseEntity<List<Estudiante>> getAllEstudiantes() {
+        List<Estudiante> estudiantes = estudianteService.getAllEstudiantes();
+        return new ResponseEntity<>(estudiantes, HttpStatus.OK);
+    }
 
-	@GetMapping("/{id}")
-	public Estudiante getEstudianteById(@PathVariable Long id) {
-		return estudianteService.getEstudianteById(id);
-	}
+    // Obtener un estudiante por ID
+    @GetMapping("/{id}")
+    public ResponseEntity<Estudiante> getEstudianteById(@PathVariable Long id) {
+        Optional<Estudiante> estudiante = estudianteService.getEstudianteById(id);
+        if(estudiante.isPresent()) {
+            return new ResponseEntity<>(estudiante.get(), HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+    }
 
-	@PostMapping
-	public Estudiante createEstudiante(@RequestBody Estudiante estudiante) {
-		return estudianteService.saveEstudiante(estudiante);
-	}
+    // Crear un nuevo estudiante
+    @PostMapping
+    public ResponseEntity<Estudiante> createEstudiante(@RequestBody Estudiante estudiante) {
+        Estudiante savedEstudiante = estudianteService.saveEstudiante(estudiante);
+        return new ResponseEntity<>(savedEstudiante, HttpStatus.CREATED);
+    }
 
-	@DeleteMapping("/{id}")
-	public void deleteEstudiante(@PathVariable Long id) {
-		estudianteService.deleteEstudiante(id);
-	}
 
+    // Actualizar un estudiante existente
+    @PutMapping("/{id}")
+    public ResponseEntity<Estudiante> updateEstudiante(@PathVariable Long id, @RequestBody Estudiante estudianteDetails) {
+        Optional<Estudiante> estudiante = estudianteService.getEstudianteById(id);
+        if(estudiante.isPresent()) {
+            Estudiante updatedEstudiante = estudiante.get();
+            updatedEstudiante.setNombre(estudianteDetails.getNombre());
+            updatedEstudiante.setApellido(estudianteDetails.getApellido());
+            updatedEstudiante.setFechaNacimiento(estudianteDetails.getFechaNacimiento());
+            updatedEstudiante.setEmail(estudianteDetails.getEmail());
+            updatedEstudiante.setCurso(estudianteDetails.getCurso());
+            updatedEstudiante.setGrado(estudianteDetails.getGrado());
+            estudianteService.saveEstudiante(updatedEstudiante);
+            return new ResponseEntity<>(updatedEstudiante, HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+    }
+
+    // Eliminar un estudiante
+    @DeleteMapping("/{id}")
+    public ResponseEntity<Void> deleteEstudiante(@PathVariable Long id) {
+        Optional<Estudiante> estudiante = estudianteService.getEstudianteById(id);
+        if(estudiante.isPresent()) {
+            estudianteService.deleteEstudiante(id);
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        } else {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+    }
 }
+
